added move patch resource product to /api/v2
