+++ src/main/resources/application-dev.properties
@@ -5,4 +5,6 @@ spring.datasource.username=rootuser
 spring.datasource.password=rootpass
 spring.h2.console.enabled=true
 spring.jpa.hibernate.ddl-auto=create-drop
+spring.servlet.multipart.max-file-size=128KB
+spring.servlet.multipart.max-request-size=128KB
 config.uploads.path=/home/nickgalan/Pictures/uploads/
